requirements.txt
slack-bolt
redis
docker
python-dotenv
requests
typer
pyright
